ShopBotSite/requirements.txt
asgiref==3.3.1
cffi==1.14.5
click==7.1.2
cryptography==3.4.6
dataclasses==0.6
Django==3.1.7
django-rest-knox==4.1.0
djangorestframework==3.12.2
djangotoolbox==1.8.0
dnspython==2.1.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
Werkzeug==1.0.1
whitenoise==5.2.0
